create single file for class header
