+++ CMakeLists.txt
@@ -22,15 +22,20 @@ add_executable(camera_inference
 add_executable(video_inference
                src/video_inference.cpp)
 
+add_executable(class_image_inference
+               src/class_image_inference.cpp)
+
 # Set include directories for all executables
 target_include_directories(image_inference PRIVATE "${ONNXRUNTIME_DIR}/include")
 target_include_directories(camera_inference PRIVATE "${ONNXRUNTIME_DIR}/include")
 target_include_directories(video_inference PRIVATE "${ONNXRUNTIME_DIR}/include")
+target_include_directories(class_image_inference PRIVATE "${ONNXRUNTIME_DIR}/include")
 
 # Link libraries for all executables
 target_link_libraries(image_inference ${OpenCV_LIBS})
 target_link_libraries(camera_inference ${OpenCV_LIBS})
 target_link_libraries(video_inference ${OpenCV_LIBS})
+target_link_libraries(class_image_inference ${OpenCV_LIBS})
 
 if(UNIX)
     message(STATUS "We are building on Linux!")
@@ -38,6 +43,7 @@ if(UNIX)
     target_link_libraries(image_inference "${ONNXRUNTIME_DIR}/lib/libonnxruntime.so")
     target_link_libraries(camera_inference "${ONNXRUNTIME_DIR}/lib/libonnxruntime.so")
     target_link_libraries(video_inference "${ONNXRUNTIME_DIR}/lib/libonnxruntime.so")
+    target_link_libraries(class_image_inference "${ONNXRUNTIME_DIR}/lib/libonnxruntime.so")
 endif(UNIX)
 
 if(APPLE)
@@ -46,6 +52,7 @@ if(APPLE)
     target_link_libraries(image_inference "${ONNXRUNTIME_DIR}/lib/libonnxruntime.dylib")
     target_link_libraries(camera_inference "${ONNXRUNTIME_DIR}/lib/libonnxruntime.dylib")
     target_link_libraries(video_inference "${ONNXRUNTIME_DIR}/lib/libonnxruntime.dylib")
+    target_link_libraries(class_image_inference "${ONNXRUNTIME_DIR}/lib/libonnxruntime.dylib")
 endif(APPLE)
 
 if(WIN32)
@@ -54,4 +61,5 @@ if(WIN32)
     target_link_libraries(image_inference "${ONNXRUNTIME_DIR}/lib/onnxruntime.lib")
     target_link_libraries(camera_inference "${ONNXRUNTIME_DIR}/lib/onnxruntime.lib")
     target_link_libraries(video_inference "${ONNXRUNTIME_DIR}/lib/onnxruntime.lib")
+    target_link_libraries(class_image_inference "${ONNXRUNTIME_DIR}/lib/onnxruntime.lib")
 endif(WIN32)
